Move the Fire Tabs promo kill-switch to an app-owned feature flag so promo code no longer imports the browser-mode impl
